requirements.txt
# ========================================
# 电商用户模块依赖配置
# ========================================

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# 数据验证
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# 安全
bcrypt==4.1.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# 配置管理
pyyaml==6.0.1

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# 工具
python-dotenv==1.0.0
